# src/openapi.yaml
openapi: "3.1.0"
info:
  title: Litix API
  version: "1.0.0"
  description: |
    API da Litix para monitoramento de processos judiciais brasileiros.
    Cadastre processos por CNJ, receba alertas de movimentacoes, gerencie clientes e webhooks.

    ## Autenticacao
    Use `Authorization: Bearer ltx_...` com sua API key.
    Crie chaves em **Configuracoes > API Keys** ou via `POST /api/v1/api-keys`.
servers:
  - url: https://app.litix.com.br/api/v1
    description: Producao
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key (ltx_...) ou token de sessao Supabase"
...
  schemas:
...
    ApiKey:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        key_prefix: { type: string }
        is_active: { type: boolean }
        last_used_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
paths:
...
  /api-keys:
    get:
      summary: Listar API keys
      tags: [API Keys]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ApiKey' } }
    post:
      summary: Criar API key
      description: A chave completa e retornada apenas uma vez nesta resposta.
      tags: [API Keys]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, default: "Default" }
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiKey' }
                  - type: object
                    properties:
                      api_key: { type: string, description: "Chave completa (exibida uma unica vez)" }
                      warning: { type: string }
  /api-keys/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: Atualizar API key
      tags: [API Keys]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                is_active: { type: boolean }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKey' }
    delete:
      summary: Revogar API key
      tags: [API Keys]
      responses:
        "204": { description: Revogada }